deployment/requirements.txt
tensorflow==2.12.0
tensorflow_hub==0.11
boto3
numpy
Pillow